decouple cost model and model optimizer in model based tuner
